spring ,camel dependency update, code refactoring

+++ src/main/java/org/entur/kingu/repository/FareZoneRepositoryImpl.java
@@ -16,29 +16,24 @@
 package org.entur.kingu.repository;
 
 
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.Query;
 import org.apache.commons.lang3.NotImplementedException;
 import org.apache.commons.lang3.StringUtils;
 import org.entur.kingu.config.ExportParams;
 import org.entur.kingu.config.VersionValidity;
+import org.entur.kingu.model.FareZone;
+import org.entur.kingu.repository.iterator.ScrollableResultIterator;
 import org.hibernate.ScrollMode;
 import org.hibernate.ScrollableResults;
 import org.hibernate.Session;
-import org.hibernate.internal.SessionImpl;
 import org.hibernate.query.NativeQuery;
-import org.entur.kingu.model.FareZone;
-import org.entur.kingu.repository.iterator.ScrollableResultIterator;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.util.Pair;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
-import java.math.BigInteger;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -221,7 +216,7 @@ public class FareZoneRepositoryImpl implements FareZoneRepositoryCustom {
 
         Session session = entityManager.unwrap(Session.class);
         NativeQuery query = session.createNativeQuery(sql);
-        return ((BigInteger) query.uniqueResult()).intValue();
+        return ((Long) query.uniqueResult()).intValue();
     }
 
     public int updateStopPlaceTariffZoneRef() {
